improve slot usage in form components + allow '.g.vue' file extension for global components
