+++ assets/components/App.g.vue
