+++ assets/components/FetchComponent.g.vue
