+++ assets/components/Form/CheckboxGroupType.g.vue
@@ -1,12 +1,20 @@
 <template>
-    <v-radio-group v-bind="Object.assign(attributes, $attrs)" v-model="form.vars.data">
-        <v-radio
+    <!-- purpose of v-radio-group is ability to pass on attributes, in particular for 'label' and 'row' -->
+    <v-radio-group v-bind="Object.assign(attributes, $attrs)">
+        <v-checkbox
             v-for="(choice, key) of choices" :key="key"
+            v-model="form.vars.data"
             :name="form.vars.full_name + '[]'"
             :label="choice.label"
             :value="choice.value"
             v-bind="choice.attr"
-        ></v-radio>
+        >
+            <!-- Pass on all slots -->
+            <slot v-if="namedSlots" v-for="slot in Object.keys(namedSlots)" :name="slot" :slot="slot"/>
+            <template v-for="slot in Object.keys(scopedSlots)" :slot="slot" slot-scope="scope">
+                <slot :name="slot" v-bind="scope"/>
+            </template>
+        </v-checkbox>
     </v-radio-group>
 </template>
 
@@ -16,7 +24,7 @@
     import IForm from "./IForm";
 
     @Component
-    export default class RadioGroupType extends Mixins(FormWidgetMixin) {
+    export default class CheckboxGroupType extends Mixins(FormWidgetMixin) {
         choices: any = [];
         created() {
             this.attributes['value'] = this.form.vars.value;
@@ -31,3 +39,11 @@
         }
     }
 </script>
+
+<style lang="scss">
+    .v-input--radio-group__input {
+        >.v-input--selection-controls {
+            margin-top: 0;
+        }
+    }
+</style>
