--- assets/components/Form/PasswordType.vue
@@ -1,12 +1,0 @@
-<template>
-    <v-text-field type="password" v-model="form.vars.data" v-bind="Object.assign(attributes, $attrs)"/>
-</template>
-
-<script lang="ts">
-    import {Component, Mixins} from 'vue-property-decorator';
-    import FormWidgetMixin from "./FormWidgetMixin.ts";
-
-    @Component
-    export default class PasswordType extends Mixins(FormWidgetMixin) {
-    }
-</script>
