+++ assets/components/Form/SwitchType.g.vue
@@ -1,5 +1,11 @@
 <template>
-    <v-switch v-model="form.vars.data" v-bind="Object.assign(attributes, $attrs)"></v-switch>
+    <v-switch v-model="form.vars.data" v-bind="Object.assign(attributes, $attrs)">
+        <!-- Pass on all slots -->
+        <slot v-if="namedSlots" v-for="slot in Object.keys(namedSlots)" :name="slot" :slot="slot"/>
+        <template v-for="slot in Object.keys(scopedSlots)" :slot="slot" slot-scope="scope">
+            <slot :name="slot" v-bind="scope"/>
+        </template>
+    </v-switch>
 </template>
 
 <script lang="ts">
