+++ assets/components/Form/TextareaType.g.vue
@@ -1,0 +1,19 @@
+<template>
+    <v-textarea v-model="form.vars.data" v-bind="Object.assign(attributes, $attrs)">
+        <!-- Pass on all slots -->
+        <slot v-if="namedSlots" v-for="slot in Object.keys(namedSlots)" :name="slot" :slot="slot"/>
+        <template v-for="slot in Object.keys(scopedSlots)" :slot="slot" slot-scope="scope">
+            <slot :name="slot" v-bind="scope"/>
+        </template>
+    </v-textarea>
+</template>
+
+<script lang="ts">
+    import {Component, Mixins} from 'vue-property-decorator';
+    import FormWidgetMixin from "./FormWidgetMixin.ts";
+
+    @Component
+    export default class TextareaType extends Mixins(FormWidgetMixin) {
+
+    }
+</script>
