+++ assets/components/MenuItem.g.vue
